# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koupii LMS API",
        "description": "API documentation for English course LMS",
        "contact": {
            "email": "support@koupii.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/classes": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Get all classes",
                "description": "Admin & Student can see all classes, Teacher can only see their own classes.",
                "operationId": "6b25f6b03203da799fad5fdc967c1457",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of classes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "20d7c432-467b-433a-b035-b43481b5ee85"
                                            },
                                            "teacher_id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "598d528e-c734-456e-b77c-7abee4cf92fa"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Theorytical Class"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "This is theory class"
                                            },
                                            "class_code": {
                                                "type": "string",
                                                "example": "lkoiun"
                                            },
                                            "cover_image": {
                                                "type": "string",
                                                "example": "cover.jpg"
                                            },
                                            "is_active": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-17T07:10:46.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-17T07:10:46.000000Z"
                                            },
                                            "teacher": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "598d528e-c734-456e-b77c-7abee4cf92fa"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Teacher User 1"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "teacher1@example.com"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "teacher"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "bio": {
                                                        "type": "string",
                                                        "example": "Teacher account 1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/classes/create": {
            "post": {
                "tags": [
                    "Classes"
                ],
                "summary": "Create a new class",
                "description": "Only admin and teacher can create a new class.",
                "operationId": "b84bd7a37320a26c569ba548ef4fb038",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Theory Class"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is theory class"
                                    },
                                    "class_code": {
                                        "type": "string",
                                        "example": "ABC12345"
                                    },
                                    "cover_image": {
                                        "description": "Upload class cover image (jpg, png)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Class created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Class created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "7c8461cf-45d7-4ce3-b2c3-dfd3760783be"
                                                },
                                                "teacher_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "598d528e-c734-456e-b77c-7abee4cf92fa"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Arabic Class"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "This is arabic class"
                                                },
                                                "class_code": {
                                                    "type": "string",
                                                    "example": "123456"
                                                },
                                                "cover_image": {
                                                    "type": "string",
                                                    "example": "cover.jpg",
                                                    "nullable": true
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-17T08:44:38.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-17T08:44:38.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Class name already exists"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/classes/delete/{id}": {
            "delete": {
                "tags": [
                    "Classes"
                ],
                "summary": "Delete a class",
                "description": "Only admin and teacher (class owner) can delete a class.",
                "operationId": "b97ca6c1832c317ea82ce5717abd9ce1",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Class ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Class not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}